Add spec test and modify classes
