+++ src/base_decorator.rb
@@ -1,0 +1,11 @@
+require_relative './nameable'
+
+class BaseDecorator
+  def initialize(nameable)
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.name
+  end
+end
