+++ src/nameable.rb
@@ -1,0 +1,5 @@
+class Nameable
+  def correct_name
+    puts 'NotImplementedError'
+  end
+end
