+++ src/person.rb
@@ -1,18 +1,19 @@
-require_relative './nameable.rb'
+require_relative './nameable'
 
 class Person < Nameable
   attr_accessor :name, :age
   attr_reader :id
 
   def initialize(age, name: 'unknown', parent_permission: true)
+    super()
     @id = Random.rand(1..1000)
-    @name = name
     @age = age
+    @name = name
     @parent_permission = parent_permission
   end
 
   def correct_name
-    return @name
+    @name
   end
 
   def can_use_services?
